feat(ios): confirm display currency change with no-conversion alert (PUL-205)
