+++ ios/Pulpe/Features/Account/CurrencySettingView.swift
@@ -25,6 +25,7 @@ struct CurrencySettingView: View {
     @State private var viewModel = CurrencySettingViewModel()
     @State private var submitSuccessTrigger = false
     @State private var isConverterExpanded = false
+    @State private var pendingCurrency: SupportedCurrency?
     @State private var saveCurrencyTask: Task<Void, Never>?
     @State private var saveSelectorToggleTask: Task<Void, Never>?
 
@@ -40,6 +41,21 @@ struct CurrencySettingView: View {
             }
             .listRowBackground(Color.surfaceContainerHigh)
             .sensoryFeedback(.success, trigger: submitSuccessTrigger)
+            // PUL-205: alert (not action sheet) confirming the flip before it
+            // persists — the copy makes explicit that amounts are not converted.
+            .alert(
+                "Changer la devise d'affichage ?",
+                isPresented: isCurrencyConfirmationPresented,
+                presenting: pendingCurrency
+            ) { currency in
+                Button("Annuler", role: .cancel) {}
+                Button("Changer") { applyCurrencyChange(to: currency) }
+            } message: { currency in
+                Text(
+                    "Tes montants existants ne sont pas convertis — 100 restera 100, "
+                        + "affiché en \(currency.rawValue). Seule la devise d'affichage change."
+                )
+            }
             .onChange(of: userSettingsStore.currency) { _, newValue in
                 viewModel.syncCurrency(newValue)
                 if isConverterExpanded {
@@ -73,16 +89,10 @@ struct CurrencySettingView: View {
                     get: { viewModel.selectedCurrency },
                     set: { newValue in
                         guard newValue != viewModel.selectedCurrency else { return }
-                        let previousCurrency = viewModel.selectedCurrency
-                        viewModel.selectedCurrency = newValue
-                        viewModel.applyConverterBase(newValue)
-                        if isConverterExpanded {
-                            viewModel.reloadRate()
-                        }
-                        saveCurrencyTask?.cancel()
-                        saveCurrencyTask = Task(name: "CurrencySetting.saveCurrency") {
-                            await persistCurrencyChange(from: previousCurrency, to: newValue)
-                        }
+                        // Stage the flip behind the PUL-205 confirmation alert.
+                        // Cancelling leaves the picker on the current currency
+                        // since this binding's `get` was never changed.
+                        pendingCurrency = newValue
                     }
                 ),
                 title: nil
@@ -130,6 +140,32 @@ struct CurrencySettingView: View {
         .accessibilityHint(
             "Active pour pouvoir entrer une dépense en EUR ou CHF, peu importe ta devise principale."
         )
+    }
+
+    // MARK: - Currency Change Confirmation (PUL-205)
+
+    /// Bridges the `SupportedCurrency?` pending state to the `Bool` binding the
+    /// alert needs: present while a value is staged, clear it on dismissal.
+    private var isCurrencyConfirmationPresented: Binding<Bool> {
+        Binding(
+            get: { pendingCurrency != nil },
+            set: { if !$0 { pendingCurrency = nil } }
+        )
+    }
+
+    private func applyCurrencyChange(to newValue: SupportedCurrency) {
+        let previousCurrency = viewModel.selectedCurrency
+        withAnimation(.snappy(duration: DesignTokens.Animation.fast)) {
+            viewModel.selectedCurrency = newValue
+        }
+        viewModel.applyConverterBase(newValue)
+        if isConverterExpanded {
+            viewModel.reloadRate()
+        }
+        saveCurrencyTask?.cancel()
+        saveCurrencyTask = Task(name: "CurrencySetting.saveCurrency") {
+            await persistCurrencyChange(from: previousCurrency, to: newValue)
+        }
     }
 
     // MARK: - Persistence
@@ -367,96 +403,6 @@ private struct ConverterRowAccessibilityModifier: ViewModifier {
             content
                 .accessibilityElement(children: .combine)
                 .accessibilityLabel(inputCombinedLabel ?? "")
-        }
-    }
-}
-
-// MARK: - ViewModel
-
-@Observable @MainActor
-final class CurrencySettingViewModel {
-    var selectedCurrency: SupportedCurrency = .chf
-    var converterInput = ""
-    var sourceCurrency: SupportedCurrency = .chf
-    var targetCurrency: SupportedCurrency = .eur
-
-    private(set) var rate: CurrencyRate?
-    private(set) var isLoadingRate = false
-    private(set) var rateFetchFailed = false
-    private var loadRateTask: Task<Void, Never>?
-
-    var convertedAmount: String {
-        let trimmed = converterInput.trimmingCharacters(in: .whitespacesAndNewlines)
-        if trimmed.isEmpty { return "—" }
-        guard let rate,
-              let inputValue = Decimal(string: trimmed.replacingOccurrences(of: ",", with: ".")) else {
-            return "—"
-        }
-        let converted = inputValue * rate.rate
-        return converted.asCurrency(targetCurrency)
-    }
-
-    var rateInfo: String? {
-        guard let rate else { return nil }
-        // PUL-114: rate is Decimal end-to-end. Format via Decimal.formatted to
-        // preserve precision (no Double bridge through %.4f).
-        let rateText = rate.rate.formatted(
-            .number.precision(.fractionLength(4)).locale(Formatters.locale(for: rate.base))
-        )
-        return "1 \(rate.base.rawValue) = \(rateText) \(rate.target.rawValue) (\(rate.date))"
-    }
-
-    /// Aligne le convertisseur : devise du compte = ligne « Depuis », l’autre devise = « Vers ».
-    /// Ne déclenche pas de réseau : la vue appelle `reloadRate()` lorsque le panneau convertisseur est ouvert.
-    func applyConverterBase(_ currency: SupportedCurrency) {
-        let newTarget: SupportedCurrency = currency == .chf ? .eur : .chf
-        // Idempotent: avoid a redundant `loadRate()` round-trip + UI flicker when the picker
-        // selection and the optimistic store update fire `applyConverterBase` back-to-back.
-        guard sourceCurrency != currency || targetCurrency != newTarget else { return }
-        sourceCurrency = currency
-        targetCurrency = newTarget
-    }
-
-    func syncCurrency(_ currency: SupportedCurrency) {
-        selectedCurrency = currency
-        applyConverterBase(currency)
-    }
-
-    func save(using store: UserSettingsStore) async {
-        await store.updateCurrency(selectedCurrency)
-    }
-
-    /// Cancels any in-flight rate fetch and starts a fresh one. Prevents stale EUR→CHF
-    /// responses from overwriting newer CHF→EUR results when the user toggles quickly.
-    func reloadRate() {
-        loadRateTask?.cancel()
-        loadRateTask = Task { [weak self] in
-            await self?.loadRate()
-        }
-    }
-
-    func loadRate() async {
-        rateFetchFailed = false
-        guard sourceCurrency != targetCurrency else {
-            rate = nil
-            return
-        }
-        isLoadingRate = true
-        defer { isLoadingRate = false }
-
-        do {
-            let fetched = try await CurrencyConversionService.shared.getRate(
-                base: sourceCurrency,
-                target: targetCurrency
-            )
-            try Task.checkCancellation()
-            rate = fetched
-            rateFetchFailed = false
-        } catch is CancellationError {
-            // Superseded by a newer request — leave state untouched.
-        } catch {
-            rate = nil
-            rateFetchFailed = true
         }
     }
 }
